Fix bug 16960 - SEGV when fetching the results from a row producing procedure
Overview
--------

The change to allow for buffering of row data fetched from IIapi_getColumns()
did not take into consideration row producing procedures. As such when a
row producing procedure is executed the driver does not calculate the row
width. When the driver comes to allocate the memory used to store the row
data from IIapi_getColumns() it multiplies the number of rows against
ii_result->rowWidth (whose value is 0), causing a SEGV in PHP.

This change extends the calculation of the row width when
IIapi_getDescriptor() indicates there is data to be fetched. As such the
driver can allocate the correct size buffer for fetching data.

git-svn-id: http://svn.php.net/repository/pecl/ingres/trunk@299790 c90b9560-bf6c-de11-be94-00142212c4b1