Fix bug 17556 - Handle errors for non-result returning statements
Overview
--------

The driver is ignoring any errors that occur for a non-select
statement executed by ingres_query() or ingres_unbuffered_query().
After executing the following SQL:
    DROP TABLE nosuchtable
the function ingres_errno() returns 0 even though the table
"nosuchtable" does not exist.

This change adds a missing call to IIapi_getQueryInfo() for
non-result returning statements. During cleanup for the request
the driver was not closing the active statement which caused
the error. This is now no longer the case.

A new test, bug_17556.phpt, has been added to test for this
problem.

git-svn-id: http://svn.php.net/repository/pecl/ingres/trunk@300165 c90b9560-bf6c-de11-be94-00142212c4b1